[BCW-57] refactor: DailyCount 인터페이스 common 폴더로 이동
