+++ be/src/main/java/com/ssb/scalendar/domain/common/projection/DailyCount.java
@@ -1,4 +1,4 @@
-package com.ssb.scalendar.domain.schedule.repository;
+package com.ssb.scalendar.domain.common.projection;
 
 import java.time.LocalDate;
